wechat_catcher/requirements.txt
wxauto4
pandas>=2.0.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
pygetwindow>=0.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
playwright>=1.40.0